Refactor and enhance converter usage
This commit represents a significant refactoring and enhancement of the converter usage within the WireMockInspector application. Key changes include the relocation of converter classes (`EnumItemsConverter`, `GraphConverter`, `StringListConverter`, `StringMatchConverter`) from the `WireMockInspector.ViewModels` namespace to a new `WireMockInspector.Converters` namespace, improving modularity and clarity. New converter classes have been added, encapsulating the conversion logic for better reusability. Updates in various XAML files reflect these namespace changes, ensuring correct converter references in data-binding expressions. Additionally, unused `using` directives have been cleaned up, and a new `SettingsWrapperTemplateSelector` class has been introduced to enhance dynamic UI generation capabilities. These changes collectively aim to improve the maintainability and functionality of the application's UI data presentation.
